Bug 1463222 - break out middleware into separate modules
..and modernize (use const, etc.)

Note that this avoids calling `res.reportError(err)` when neither
req.authorize nor res.reply were called, and instead just reports the
event to sentry. The existing approach was causing unhandled promise
rejections that were inexplicably not detected before the refactor.
